refactor: split UniversalValidationRules.ts into per-rule modules
- Created 15 individual rule files (54-111 lines each)
- Extracted shared helper to helpers.ts (27 lines)
- Converted UniversalValidationRules.ts to barrel (53 lines)
- All files under FILE_MODULARITY_SPEC thresholds
- Zero consumer changes (barrel pattern)
